+++ SkillModifiers/phantom_impact.lua
@@ -2,7 +2,7 @@ local scale = 1.25
 local phantom_impact = SkillModifierManager.register_modifier("phantom_impact")
 phantom_impact:set_add_func(function(data, modifier_index)
     local actor = data.skill.parent
-    if actor.is_local == 0 then
+    if not gm.bool(actor.is_local) then
         return
     end
     Instance_ext.add_skill_bullet_captrue_local(actor, data.skill.slot_index, data:get_id(modifier_index),
@@ -13,7 +13,7 @@ phantom_impact:set_add_func(function(data, modifier_index)
 end)
 phantom_impact:set_remove_func(function(data, modifier_index)
     local actor = data.skill.parent
-    if actor.is_local == 0 then
+    if not gm.bool(actor.is_local) then
         return
     end
     Instance_ext.remove_skill_instance_captrue(actor, data.skill.slot_index, data:get_id(modifier_index))