+++ SkillModifiers/void_power.lua
@@ -4,7 +4,7 @@ void_power:set_add_func(function(data, modifier_index)
         return origin_value * 2 ^ Utils.get_empty_skill_num(data.skill.parent)
     end)
     local actor = data.skill.parent
-    if actor.is_local == 0 then
+    if not gm.bool(actor.is_local) then
         return
     end
     Instance_ext.add_callback(data.skill.parent, "post_local_drop", data:get_id(), function()
@@ -16,7 +16,7 @@ void_power:set_add_func(function(data, modifier_index)
 end)
 void_power:set_remove_func(function(data, modifier_index)
     local actor = data.skill.parent
-    if actor.is_local == 0 then
+    if not gm.bool(actor.is_local) then
         return
     end
     Instance_ext.remove_callback(data.skill.parent, "post_local_drop", data:get_id())